slack_sample/requirements.txt
slackclient
slack_sdk
